==> Code/2_WGCNA.Rmd <==
---
title: "WGCNA clustering"
author: "Wei Liu"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r packages, echo=F, message=F}
rm(list=ls())
library("WGCNA")
library("openxlsx")
library("xlsx")
# samples preparation
library("dplyr")
#
library("biomaRt")
library("tximport")
library("DESeq2")

# elastic net
library("glmnet")
library("clusterProfiler")
# heatmap
library("ComplexHeatmap")
library("RColorBrewer")
library("ggplot2")
options(stringsAsFactors = FALSE)

```



### Treatment sample table
```{r}
load("~/Dropbox/sheep_arc/data/res/DESeq2_norm.RData")
load("~/Dropbox/sheep_arc/data/res/sample_table_w1_7.RData")
week_w <- "Week2"
sample_table <- subset(subset_samples, (group=="Group1"|group=="Group3")& week==week_w)
DESeq2_norm_1vs3 <- DESeq2_norm[,match(sample_table$sample,colnames(DESeq2_norm))]

```

### Start WGCNA
#### Sample tree clusters
```{r}
  input_matrix <- t(DESeq2_norm_1vs3)
  gsg = goodSamplesGenes(input_matrix, verbose = 3)
  gsg$allOK
  sampleTree = hclust(dist(input_matrix), method = "average")
  sizeGrWindow(12,9)
  par(cex = 0.6)
  par(mar = c(0,4,2,0))
  plot(sampleTree, main ="WGCNA sample clustering", sub="", xlab="", cex.lab = 1.5,
       cex.axis = 1.5, cex.main = 2, family="Helvetica")
```

#### calculate rank power to do clustering
```{r}
  powers = c(c(1:10), seq(from = 12, to=40, by=2))
  sft = pickSoftThreshold(input_matrix, powerVector = powers, verbose = 5)

  pdf(paste0("~/Dropbox/sheep_arc/WGCNA_control/soft_threshold_g1g3_",week_w,".pdf"), width =9, height = 5)
  #sizeGrWindow(9, 5)
  par(mfrow = c(1,2))
  cex1 = 0.85
  # Scale-free topology fit index as a function of the soft-thresholding power
  plot(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
       xlab="Soft Threshold (power)",ylab="Scale Free Topology Model Fit,signed R^2",type="n",main = "Scale independence ")
  text(sft$fitIndices[,1],-sign(sft$fitIndices[,3])*sft$fitIndices[,2],labels=powers,cex=cex1,col="red")
  abline(h=0.9,col="red")
  # Mean connectivity as a function of the soft-thresholding power
  plot(sft$fitIndices[,1], sft$fitIndices[,5],
       xlab="Soft Threshold (power)",ylab="Mean Connectivity", type="n",
       main = paste("Mean connectivity"))
  text(sft$fitIndices[,1], sft$fitIndices[,5], labels=powers, cex=cex1,col="red")
  dev.off()
```


#### write files
```{r}
MEList = moduleEigengenes(input_matrix, colors = clusters,align = "along average", scale=T)
eigen <- MEList$eigengenes
rownames(eigen) <- rownames(input_matrix)
write.csv(eigen, file =paste0("./WGCNA_res/eigengenes_clusters_g1g3_",week_w,".csv"), row.names = T, col.names = T)
#write.csv(eigen, file =paste0("WGCNA_control/eigengenes_clusters_g2g4_",week_w,".csv"), row.names = T, col.names = T)
output_matrix <- cbind(t(input_matrix), clusters)
#write.csv(output_matrix, file=paste0("WGCNA_control/DESeq_norm_genematrix_g2g4_",week_w,".csv"), row.names = T, col.names = T)
write.csv(output_matrix, file=paste0("./WGCNA_res/DESeq_norm_genematrix_g1g3_",week_w,".csv"), row.names = T, col.names = T)

```


```{r, elastic net with worm burdens and FEC}
week_w <- "Week4"
# READ Gene and Eigen TABLES
DESeq2_gene_matrix <- read.csv(paste0("~/Dropbox/sheep_arc/WGCNA/DESeq_norm_genematrix_g1g3_",week_w,".csv"), header = T)
rownames(DESeq2_gene_matrix) <- DESeq2_gene_matrix[,1]
DESeq2_gene_matrix <- DESeq2_gene_matrix[,-1]

eigen <- read.csv(paste0("~/Dropbox/sheep_arc/WGCNA/eigengenes_clusters_g1g3_",week_w,".csv"), header = T)
rownames(eigen) <- eigen[,1]
eigen <- eigen[,-1]
eigen <- subset(eigen, select=-c(MEgrey))

# Worm and FEC

 Worm <- read.csv("~/Dropbox/sheep_arc/data/res/datTraits_worm_13.csv", header = T)
 Worm <- Worm[,-1]
 Worm_biopsy <- subset(Worm, Treatment == "Immunised Biopsy")
 Worm_biopsy$Total.Worms[ Worm_biopsy$Total.Worms==0] <- 1

 FEC <- read.csv("~/Dropbox/sheep_arc/data/res/datTraits_AUC_13.csv", header = T)
 FEC <- FEC[,-1]
 FEC_biopsy <- subset(FEC, Treatment == "Immunised Biopsy")

identical(FEC_biopsy$Animal.ID, Worm_biopsy$Animal.ID)
eigen_order <- eigen[match(as.character(Worm_biopsy$Animal.ID), substr(rownames(eigen),1,4)),]
identical(as.character(Worm_biopsy$Animal.ID), substr(rownames(eigen_order),1,4))

fit_elasticnet <- cv.glmnet(scale(as.matrix(eigen_order)), Worm_biopsy$Total.Worms, family = "poisson", alpha=0) # w1-7 except 3: alpha 0.01. week3: 0 ridge regression
plot(fit_elasticnet)
fit_elasticnet$lambda.min
para_elasticnet <- as.data.frame(as.matrix(predict(fit_elasticnet,type="coef", s="lambda.min")))

para_df <- para_elasticnet
#G1 G3 week3
length(which(abs(para_df$`1`)!=0))
para_df

para_df$gene<- rownames(para_df)
colnames(para_df) <- c("Estimate", "gene")
para_df <- para_df[-1,]
#week3
para_df_sub <- subset(para_df, abs(Estimate)!=0)

MtoH <-colorRampPalette(c('yellow' ,'firebrick2'))
LtoM <-colorRampPalette(c('blue','chartreuse'))
library(RColorBrewer)
library(scales)
ggplot(data=para_df_sub, aes(x= reorder(gene,-Estimate), y = Estimate, fill=Estimate))+
  geom_bar(stat = 'identity',color="black") +
  scale_fill_distiller(palette = "Spectral")+
  labs(title="WGCNA clusters ranked by elastic net: Worm Count ~ Clusters ", x="Gene Clusters", y = "coeff.mean")+
  coord_flip()+
  theme_minimal()
#ggsave(paste0("WGCNA/elasticnet_coeff_nonzero_",week_w,".pdf"), width = 10, height = 10)

```

```{r, KEGG enrichment}
corre <- "down"

gene_table <- read.csv("~/Dropbox/sheep_arc/data/read_in/gene_table_18k.csv", sep="\t")
colnames(gene_table) <- c("id", "oas")
gene_id <- match(DESeq2_gene_matrix$clusters,substring(para_df_sub$gene,3))
match_table <- as.data.frame(na.omit(DESeq2_gene_matrix[which(!is.na(gene_id)),]))
res_table <- as.data.frame(cbind(rownames(match_table), match_table$clusters))
colnames(res_table) <- c("gene","cl_id")
res_coef.mean <- para_df_sub[match(res_table$cl_id, substring(para_df_sub$gene,3)), "Estimate"]

table_oas_id <- cbind(gene_table[match(res_table$gene, gene_table$id),], res_table$cl_id, res_coef.mean)
  table_oas_id <- na.omit(table_oas_id)
  table_oas_id$oas <- substring(table_oas_id$oas,5)

#week3

table_oas_up <- subset(table_oas_id, res_coef.mean>0)
table_oas_down <- subset(table_oas_id, res_coef.mean<0)


write.table(table_oas_up$id, file=paste0("WGCNA/all_gene_list_beforeKEGG_up_",week_w,".csv"), row.names = F, col.names = F, quote = F)
write.table(table_oas_down$id, file=paste0("WGCNA/all_gene_list_beforeKEGG_down_",week_w,".csv"),row.names = F, col.names = F, quote = F)


if (corre == "up") {
  table_split <- split(table_oas_up , f =table_oas_up$`res_table$cl_id`)
} else if (corre == "down") {
table_split <- split(table_oas_down , f =table_oas_down$`res_table$cl_id`)
}


oas_cluster <- lapply(table_split, function(x) x[,2])

ck <- compareCluster(geneCluster = oas_cluster,
                     fun = "enrichKEGG",
                     organism="oas",  pvalueCutoff=0.05)
pdf(paste0("WGCNA/KEGG_g1g3_",week_w,"_",corre,".pdf"),width=20,height=10)
dotplot(ck, colorBy = "p.adjust", showCategory = 20, by = "geneRatio",
        includeAll = TRUE, font.size = 12)
dev.off()
# save gene list
kegg_result <- ck@compareClusterResult
kegg_list <- strsplit(kegg_result$geneID, "/")
geneID_mat <- as.data.frame(str_split_fixed(kegg_result$geneID, "/",max(lengths(kegg_list))))
gene_name_mat <- NULL
for (i in 1:max(lengths(kegg_list))) {
  Gene_Name <- table_oas_id[match(geneID_mat[,i], table_oas_id$oas),"id"]
  gene_name_mat <- cbind(Gene_Name,gene_name_mat)
}
kegg_result <- cbind(kegg_result, gene_name_mat)
write.csv(kegg_result,file=paste0("WGCNA/genelist_KEGG_g1g3_",week_w,"_",corre,".csv"), row.names = F, col.names = T)

```

```{r, correlation with Worm and FEC}
week_w = "Week2"
corre = "down"

kegg_result <- read.csv(paste0("~/Dropbox/sheep_arc/WGCNA/genelist_KEGG_g1g3_",week_w,"_",corre,".csv"))

#WEEK1
#kegg_result_sub <- subset(kegg_result, Cluster=="black"|Cluster=="coral2"|Cluster=="darkred"|Cluster=="greenyellow") # UP

#WEEK2
kegg_result_sub <- subset(kegg_result, Cluster=="blueviolet") # DOWN

#WEEK3
#kegg_result_sub <- subset(kegg_result, Cluster=="brown"|Cluster=="lemonchiffon3"|Cluster=="lightgoldenrod"|Cluster=="yellow")

#WEEK4
#kegg_result_sub <- subset(kegg_result, Cluster=="blueviolet") # down
#kegg_result_sub <- subset(kegg_result, Cluster=="lightblue1"|Cluster=="saddlebrown") # UP

#WEEK5
#kegg_result_sub <- subset(kegg_result, Cluster=="deepskyblue"|Cluster=="peru") # down
#kegg_result_sub <- subset(kegg_result, Cluster=="darkorange"|Cluster=="lightgreen"|Cluster=="magenta2"|Cluster=="lightsteelblue1"|Cluster=="skyblue") # UP

#WEEK7
#kegg_result_sub <- subset(kegg_result, Cluster=="lightyellow") # down
#kegg_result_sub <- subset(kegg_result, Cluster=="mediumpurple3"|Cluster=="black"|Cluster=="honeydew1"|Cluster=="yellowgreen") # UP



kegg_gene_only <- kegg_result_sub[,grep("Gene_Name", colnames(kegg_result_sub))]
kegg_gene_unique<- as.character(unique(na.omit(unlist(kegg_gene_only))))

load("data/res/DESeq2_norm.RData")
load("data/res/sample_table_w1_7.RData")

sample_table <- subset(subset_samples, week==week_w)
kegg_gene_exp<- DESeq2_norm[match(kegg_gene_unique,rownames(DESeq2_norm)),match(sample_table$sample,colnames(DESeq2_norm))]
colnames(kegg_gene_exp) <- paste(substring(colnames(kegg_gene_exp),6,6),colnames(kegg_gene_exp), sep="_")


kegg_gene_scale <- t(scale(t(kegg_gene_exp)))
kegg_gene_order <- kegg_gene_scale[,order(colnames(kegg_gene_scale))]

gg_color_hue <- function(n) {
    hues = seq(15, 375, length=n+1)
    hcl(h=hues, l=65, c=100)[1:n]
}
colorVec = gg_color_hue(4)
names(colorVec) = 1:4
weiha = HeatmapAnnotation(df = data.frame(group=substr(colnames(kegg_gene_order),1,1)),col = list(group= colorVec))
png(file = paste0("WGCNA/complexhm_",week_w,"_",corre,".png"), height = 18, width = 18, units = 'in', res = 600)
Heatmap(kegg_gene_order,  name = "", top_annotation = weiha, cluster_columns = F)
dev.off()



fit_multi <- cv.glmnet(t(kegg_gene_order),
                         substr(colnames(kegg_gene_order),1,1),
                         family = "multinomial",  type.multinomial = "grouped",alpha=0)
fit_coef <- coef(fit_multi,  s = "lambda.min")

# heatmap of coefficient
print(nrow(kegg_gene_order))
para_mat <- matrix(0,nrow(kegg_gene_order),4)
rownames(para_mat) <- rownames(fit_coef[[1]])[-1]
top_g <- NULL
top_gene <- NULL
for (i in 1:4 ) {
  para_mat[,i]<- as.data.frame(as.numeric(fit_coef[[i]]))[-1,]
  top_g<- order(para_mat[,i])#[c(1:5,66:70)]
  top_gene <- c(top_g,top_gene)
}
top_gene <- unique(top_gene)
#para_mat <- para_mat[top_gene,]

colnames(para_mat) <- c("G1", "G2", "G3", "G4")
longData <- melt(para_mat)
head(longData, 20)
max(para_mat)

myPalette <- colorRampPalette(rev(brewer.pal(11, "Spectral")), space="Lab")
ggplot(longData,aes(x = Var2, y = Var1, fill = value))+
  geom_tile()+
  scale_fill_gradientn(colours = myPalette(100),name="GLM Coeff",limits = c(min(para_mat),max(para_mat)),oob=squish)+
  scale_x_discrete(expand = c(0, 0))+
  scale_y_discrete(expand = c(0, 0))+
  coord_equal()+
  theme_bw()+
  theme(text = element_text(size=14), axis.text.y = element_text(size=12),
        axis.text.x = element_text(angle = 45, vjust = 1, size = 12, hjust = 1),
        axis.title.y = element_blank(), axis.title.x = element_blank())+
coord_flip()
ggsave(paste0("WGCNA/multinomial_coeff_nonzero_",week_w,"_",corre,".png"), width = 4, height = 2)


# correlation with worm
Worm <- read.csv("data/res/datTraits_worm.csv", header = T)
Worm <- Worm[,-1]
Worm_biopsy <- subset(Worm, Treatment == "Immunised Biopsy"|Treatment=="Control Biopsy")
Worm_biopsy$Total.Worms[ Worm_biopsy$Total.Worms==0] <- 1
Worm_biopsy <- Worm_biopsy[match(substring(colnames(kegg_gene_order),3,6), Worm_biopsy$Animal.ID),]


for (GENE in rownames(para_mat)) {
gene_df <- as.data.frame(cbind(kegg_gene_order[GENE,], log2(Worm_biopsy$Total.Worms)))
colnames(gene_df) <- c("gene","worm")
ggplot(gene_df, aes(x=gene, y=worm)) +
geom_point()+
geom_smooth(method=lm)+
theme_bw()
ggsave(paste0("WGCNA/gene_worm_correlation/correlation_",week_w,"_",corre,"_",GENE,".png"), height = 6, width = 6)
}





```



```{r, ingenuity pathway analysis}

week_w <- "Week5"
corre <- "down"
read_ipa_table <- read.xlsx2(paste0("WGCNA/ingenuity_res/all_gene_list_beforeKEGG_",corre,"_",week_w,".xls"),1,header = T,colIndex=1:5)
ipa_table <- as.data.frame(read_ipa_table[-1,])
colnames(ipa_table) <- read_ipa_table[1,]
ipa_table$p_value <- 1/(10^as.numeric(ipa_table$` -log(p-value)`))
ipa_table_pvalue <- subset(ipa_table, p_value <= 0.01)
split_gene <- strsplit(ipa_table_pvalue$Molecules, ",")

load("data/res/DESeq2_norm.RData")
load("data/res/sample_table_w1_7.RData")

kegg_gene_unique <- split_gene[[16]]
sample_table <- subset(subset_samples, week==week_w)
kegg_gene_exp<- DESeq2_norm[match(kegg_gene_unique,rownames(DESeq2_norm)),match(sample_table$sample,colnames(DESeq2_norm))]
colnames(kegg_gene_exp) <- paste(substring(colnames(kegg_gene_exp),6,6),colnames(kegg_gene_exp), sep="_")


kegg_gene_scale <- t(scale(t(kegg_gene_exp)))
kegg_gene_order <- kegg_gene_scale[,order(colnames(kegg_gene_scale))]

# correlation with worm
Worm <- read.csv("data/res/datTraits_worm.csv", header = T)
Worm <- Worm[,-1]
Worm_biopsy <- subset(Worm, Treatment == "Immunised Biopsy"|Treatment=="Control Biopsy")
Worm_biopsy$Total.Worms[ Worm_biopsy$Total.Worms==0] <- 1
Worm_biopsy <- Worm_biopsy[match(substring(colnames(kegg_gene_order),3,6), Worm_biopsy$Animal.ID),]

kegg_gene_order <- kegg_gene_order[,match(substring(colnames(kegg_gene_order),3,6), Worm_biopsy$Animal.ID)]
cor_mat <- vector()
p_mat <- vector()
for (i in 1:length(kegg_gene_unique)) {
gene_df <- as.data.frame(cbind(kegg_gene_order[kegg_gene_unique[i],], log2(Worm_biopsy$Total.Worms)))
colnames(gene_df) <- c("gene","worm")
ggplot(gene_df, aes(x=gene, y=worm)) +
geom_point()+
geom_smooth(method=lm)+
theme_bw()
ggsave(paste0("WGCNA/ingenuity_res/gene_worm_correlation/correlation_",week_w,"_",corre,"_",kegg_gene_unique[i],".png"), height = 6, width = 6)
p_mat[i] <- cor.test(kegg_gene_order[kegg_gene_unique[i],],log2(Worm_biopsy$Total.Worms), alternative = "two.sided")[[3]]
cor_mat[i]<- cor.test(kegg_gene_order[kegg_gene_unique[i],],log2(Worm_biopsy$Total.Worms), alternative = "two.sided")[[4]]
}

CORR_MAT <- as.data.frame(cbind(cor_mat, p_mat))
rownames(CORR_MAT) <- kegg_gene_unique
length(which(CORR_MAT$cor_mat<=0))/length(kegg_gene_unique)
test <- subset(CORR_MAT, cor_mat <=0 & p_mat<=0.05)

```


```{r, response letter ingenuity pathway }

gene_deseq_table <- read.csv("./WGCNA_res/DESeq_norm_genematrix_g1g3_Week4.csv")

clusters_enet_coff <- c("snow2", "lightcyan1","violet", "springgreen2", "plum", "sienna2", "lavenderblush1", "royalblue4", "ivory4", "deepskyblue1", "salmon3", "bisque2", "lightslateblue", "honeydew4", "coral3")

for (i in 1:length(clusters_enet_coff)) {
gene_subset_table <- subset(gene_deseq_table, clusters==clusters_enet_coff[i])$X
write.table(gene_subset_table, file=paste0("./WGCNA_res/ingenuity/gene_list_",clusters_enet_coff[i],".csv"), row.names = F, col.names=F, quote = F)
}
subset(gene_deseq_table, X=="WNT6")$clusters

read.table("")

```
